derive: implement a 'prefix:...' casing style
This prepends a string to all arg names in a Command. It's also a recursive style: it can be followed up with another casing style as well (it defaults to Snake) by typing 'prefix:foo,kebab'.

To use, put an attr on a struct like:
```rust
 #[clap(rename_all = "prefix:foobar")]
 struct Foobar {
     #[clap(long)]
     hello_world: String,
 }
```

This would result in an arg `--foobar.hello-world`.